+++ crates/vertigo/src/driver_module/src_js/jsvalue_types.ts
@@ -34,6 +34,15 @@ export type JsValueType
     | { type: typeof JsValueConst.Object, value: JsValueMapType }
     | { type: typeof JsValueConst.Json, value: JsJsonType };
 
-    interface JsValueMapType {
-        [key: string]: JsValueType
-    }
+export interface JsValueMapType {
+    [key: string]: JsValueType
+}
+
+export function isJsObject(value: JsValueType): value is { type: typeof JsValueConst.Object, value: JsValueMapType } {
+    return (
+        value !== null &&
+        typeof value === 'object' &&
+        'type' in value &&
+        value.type === JsValueConst.Object
+    );
+}